# feat: use assign_regions api to speed up the witness assignment