--- a/src/hash.rs
+++ b/src/hash.rs
@@ -475,10 +475,6 @@ impl<'d, F: Hashable, const STEP: usize, PC: PermuteChip<F, F::SpecType, 3, 2>>
         let config = &self.config;
         let data = self.data;
 
-        let mut states_in = Vec::new();
-        let mut states_out = Vec::new();
-        let hash_helper = F::hasher();
-
         let inputs_i = data
             .inputs
             .iter()
@@ -499,136 +495,153 @@ impl<'d, F: Hashable, const STEP: usize, PC: PermuteChip<F, F::SpecType, 3, 2>>
             .chain(std::iter::repeat(None))
             .take(self.calcs);
 
-        let mut is_new_sponge = true;
-        let mut process_start = 0;
-        let mut state: [F; 3] = [F::zero(); 3];
-        let mut last_offset = 0;
+        let data: Vec<(usize, ((Option<&[F; 2]>, Option<&u64>), Option<&F>))> =
+            inputs_i.zip(controls_i).zip(checks_i).enumerate().collect();
+        let data_len = data.len();
+        let ret = data
+            .group_by(|(_, ((_, control), _)), _| control.copied().unwrap_or(0) > STEP as u64)
+            .map(|data| -> Result<PermutedStatePair<PC::Word>, Error> {
+                let mut is_new_sponge = true;
+                let mut states_in = Vec::new();
+                let mut states_out = Vec::new();
+                let hash_helper = F::hasher();
+                let mut process_start = 0;
+                let mut state = [F::zero(); 3];
 
-        for (i, ((inp, control), check)) in inputs_i.zip(controls_i).zip(checks_i).enumerate() {
-            let control = control.copied().unwrap_or(0);
-            let offset = i + begin_offset;
-            last_offset = offset;
+                for (i, ((inp, control), check)) in data.iter() {
+                    let control = control.copied().unwrap_or(0u64);
+                    let offset = i + begin_offset;
 
-            let control_as_flag = F::from_u128(control as u128 * HASHABLE_DOMAIN_SPEC);
+                    let control_as_flag = F::from_u128(control as u128 * HASHABLE_DOMAIN_SPEC);
 
-            if is_new_sponge {
-                state[0] = control_as_flag;
-                process_start = offset;
-            }
+                    if is_new_sponge {
+                        state[0] = control_as_flag;
+                        process_start = offset;
+                    }
 
-            let inp = inp
-                .map(|[a, b]| [*a, *b])
-                .unwrap_or_else(|| [F::zero(), F::zero()]);
+                    let inp = inp
+                        .map(|[a, b]| [*a, *b])
+                        .unwrap_or_else(|| [F::zero(), F::zero()]);
 
-            state.iter_mut().skip(1).zip(inp).for_each(|(s, inp)| {
-                if is_new_sponge {
-                    *s = inp;
-                } else {
-                    *s += inp;
+                    state.iter_mut().skip(1).zip(inp).for_each(|(s, inp)| {
+                        if is_new_sponge {
+                            *s = inp;
+                        } else {
+                            *s += inp;
+                        }
+                    });
+
+                    let state_start = state;
+                    hash_helper.permute(&mut state); //here we calculate the hash
+
+                    //and sanity check ...
+                    if let Some(ck) = check {
+                        assert_eq!(
+                            **ck, state[0],
+                            "hash output not match with expected at {offset}"
+                        );
+                    }
+
+                    let current_hash = state[0];
+
+                    //assignment ...
+                    region.assign_fixed(
+                        || "assign q_enable",
+                        self.config.q_enable,
+                        offset,
+                        || Value::known(F::one()),
+                    )?;
+
+                    let c_start = [0; 3]
+                        .into_iter()
+                        .enumerate()
+                        .map(|(i, _)| {
+                            region.assign_advice(
+                                || format!("state input {i}_{offset}"),
+                                config.hash_table_aux[i],
+                                offset,
+                                || Value::known(state_start[i]),
+                            )
+                        })
+                        .collect::<Result<Vec<_>, _>>()?;
+
+                    let c_end = [5, 3, 4]
+                        .into_iter()
+                        .enumerate()
+                        .map(|(i, j)| {
+                            region.assign_advice(
+                                || format!("state output {i}_{offset}"),
+                                config.hash_table_aux[j],
+                                offset,
+                                || Value::known(state[i]),
+                            )
+                        })
+                        .collect::<Result<Vec<_>, _>>()?;
+
+                    for (tip, col, val) in [
+                        ("hash input first", config.hash_table[1], inp[0]),
+                        ("hash input second", config.hash_table[2], inp[1]),
+                        ("state input control", config.hash_table[3], control_as_flag),
+                        (
+                            "state beginning flag",
+                            config.hash_table[4],
+                            if is_new_sponge { F::one() } else { F::zero() },
+                        ),
+                        (
+                            "state input control_aux",
+                            config.control_aux,
+                            control_as_flag.invert().unwrap_or_else(F::zero),
+                        ),
+                        (
+                            "state continue control",
+                            config.s_sponge_continue,
+                            if is_new_sponge { F::zero() } else { F::one() },
+                        ),
+                    ] {
+                        region.assign_advice(
+                            || format!("{tip}_{offset}"),
+                            col,
+                            offset,
+                            || Value::known(val),
+                        )?;
+                    }
+
+                    is_new_sponge = control <= STEP as u64;
+
+                    //fill all the hash_table[0] with result hash
+                    if is_new_sponge {
+                        (process_start..=offset).try_for_each(|ith| {
+                            region
+                                .assign_advice(
+                                    || format!("hash index_{ith}"),
+                                    config.hash_table[0],
+                                    ith,
+                                    || Value::known(current_hash),
+                                )
+                                .map(|_| ())
+                        })?;
+                    }
+
+                    //we directly specify the init state of permutation
+                    let c_start_arr: [_; 3] = c_start.try_into().expect("same size");
+                    states_in.push(c_start_arr.map(PC::Word::from));
+                    let c_end_arr: [_; 3] = c_end.try_into().expect("same size");
+                    states_out.push(c_end_arr.map(PC::Word::from));
                 }
-            });
+                Ok((states_in, states_out))
+            })
+            .collect::<Result<Vec<_>, _>>()?;
 
-            let state_start = state;
-            hash_helper.permute(&mut state); //here we calculate the hash
-
-            //and sanity check ...
-            if let Some(ck) = check {
-                assert_eq!(
-                    *ck, state[0],
-                    "hash output not match with expected at {offset}"
-                );
-            }
-
-            let current_hash = state[0];
-
-            //assignment ...
-            region.assign_fixed(
-                || "assign q_enable",
-                self.config.q_enable,
-                offset,
-                || Value::known(F::one()),
-            )?;
-
-            let c_start = [0; 3]
-                .into_iter()
-                .enumerate()
-                .map(|(i, _)| {
-                    region.assign_advice(
-                        || format!("state input {i}_{offset}"),
-                        config.hash_table_aux[i],
-                        offset,
-                        || Value::known(state_start[i]),
-                    )
-                })
-                .collect::<Result<Vec<_>, _>>()?;
-
-            let c_end = [5, 3, 4]
-                .into_iter()
-                .enumerate()
-                .map(|(i, j)| {
-                    region.assign_advice(
-                        || format!("state output {i}_{offset}"),
-                        config.hash_table_aux[j],
-                        offset,
-                        || Value::known(state[i]),
-                    )
-                })
-                .collect::<Result<Vec<_>, _>>()?;
-
-            for (tip, col, val) in [
-                ("hash input first", config.hash_table[1], inp[0]),
-                ("hash input second", config.hash_table[2], inp[1]),
-                ("state input control", config.hash_table[3], control_as_flag),
-                (
-                    "state beginning flag",
-                    config.hash_table[4],
-                    if is_new_sponge { F::one() } else { F::zero() },
-                ),
-                (
-                    "state input control_aux",
-                    config.control_aux,
-                    control_as_flag.invert().unwrap_or_else(F::zero),
-                ),
-                (
-                    "state continue control",
-                    config.s_sponge_continue,
-                    if is_new_sponge { F::zero() } else { F::one() },
-                ),
-            ] {
-                region.assign_advice(
-                    || format!("{tip}_{offset}"),
-                    col,
-                    offset,
-                    || Value::known(val),
-                )?;
-            }
-
-            is_new_sponge = control <= STEP as u64;
-
-            //fill all the hash_table[0] with result hash
-            if is_new_sponge {
-                (process_start..=offset).try_for_each(|ith| {
-                    region
-                        .assign_advice(
-                            || format!("hash index_{ith}"),
-                            config.hash_table[0],
-                            ith,
-                            || Value::known(current_hash),
-                        )
-                        .map(|_| ())
-                })?;
-            }
-
-            //we directly specify the init state of permutation
-            let c_start_arr: [_; 3] = c_start.try_into().expect("same size");
-            states_in.push(c_start_arr.map(PC::Word::from));
-            let c_end_arr: [_; 3] = c_end.try_into().expect("same size");
-            states_out.push(c_end_arr.map(PC::Word::from));
+        let mut states_in = vec![];
+        let mut states_out = vec![];
+        for (s_in, s_out) in ret.into_iter() {
+            states_in.extend(s_in);
+            states_out.extend(s_out);
         }
 
         // set the last row is "custom", a row both enabled and customed
         // can only fill a padding row ([0, 0] in MPT mode)
-        config.s_custom.enable(region, last_offset)?;
+        config.s_custom.enable(region, data_len + begin_offset)?;
         Ok((states_in, states_out))
     }
 
@@ -652,15 +665,14 @@ impl<'d, F: Hashable, const STEP: usize, PC: PermuteChip<F, F::SpecType, 3, 2>>
             },
         )?;
 
-        layouter.assign_region(|| "hash table custom", |mut region| {
-            self.fill_hash_tbl_custom(&mut region)
-        })?;
+        layouter.assign_region(
+            || "hash table custom",
+            |mut region| self.fill_hash_tbl_custom(&mut region),
+        )?;
 
         let (states_in, states_out) = layouter.assign_region(
             || "hash table",
-            |mut region| {
-                self.fill_hash_tbl_body(&mut region, 0)
-            },
+            |mut region| self.fill_hash_tbl_body(&mut region, 0),
         )?;
 
         let mut chip_finals = Vec::new();
@@ -674,20 +686,22 @@ impl<'d, F: Hashable, const STEP: usize, PC: PermuteChip<F, F::SpecType, 3, 2>>
             chip_finals.push(final_state);
         }
 
-        layouter.assign_region(
-            || "final state dummy",
-            |mut region| {
-                for (state, final_state) in states_out.iter().zip(chip_finals.iter()) {
-                    for (s_cell, final_cell) in state.iter().zip(final_state.iter()) {
-                        let s_cell: AssignedCell<F, F> = s_cell.clone().into();
-                        let final_cell: AssignedCell<F, F> = final_cell.clone().into();
-                        region.constrain_equal(s_cell.cell(), final_cell.cell())?;
-                    }
+        let assignments = states_out
+            .iter()
+            .flatten()
+            .zip(chip_finals.iter().flatten())
+            .map(|(s_cell, final_cell)| {
+                |mut region: Region<'_, F>| -> Result<(), Error> {
+                    let s_cell: AssignedCell<F, F> = s_cell.clone().into();
+                    let final_cell: AssignedCell<F, F> = final_cell.clone().into();
+                    region.constrain_equal(s_cell.cell(), final_cell.cell())?;
+                    Ok(())
                 }
+            })
+            .collect();
 
-                Ok(())
-            },
-        )
+        layouter.assign_regions(|| "final state dummy", assignments)?;
+        Ok(())
     }
 }
 
@@ -714,12 +728,14 @@ mod tests {
     use super::*;
     use halo2_proofs::halo2curves::bn256::{Bn256, G1Affine};
     use halo2_proofs::halo2curves::group::ff::PrimeField;
-    use halo2_proofs::plonk::{keygen_vk, keygen_pk, create_proof, verify_proof};
+    use halo2_proofs::plonk::{create_proof, keygen_pk, keygen_vk, verify_proof};
     use halo2_proofs::poly::commitment::ParamsProver;
-    use halo2_proofs::poly::kzg::commitment::{ParamsKZG, ParamsVerifierKZG, KZGCommitmentScheme};
+    use halo2_proofs::poly::kzg::commitment::{KZGCommitmentScheme, ParamsKZG, ParamsVerifierKZG};
     use halo2_proofs::poly::kzg::multiopen::{ProverSHPLONK, VerifierSHPLONK};
     use halo2_proofs::poly::kzg::strategy::SingleStrategy;
-    use halo2_proofs::transcript::{Blake2bWrite, Challenge255, TranscriptWriterBuffer, Blake2bRead, TranscriptReadBuffer};
+    use halo2_proofs::transcript::{
+        Blake2bRead, Blake2bWrite, Challenge255, TranscriptReadBuffer, TranscriptWriterBuffer,
+    };
     use halo2_proofs::{circuit::SimpleFloorPlanner, plonk::Circuit};
 
     #[test]