--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,6 +1,7 @@
 //! mpt demo circuits
 //
 
+#![feature(slice_group_by)]
 #![allow(dead_code)]
 #![allow(unused_macros)]
 #![deny(missing_docs)]