See #23 
===
### Bechmarking

For each phase (in AMD EPYC 7702 64-Core Processor x 128 threads):

before
```
ASSIGNMENT_TYPE: default Elapsed: 22.257506277s
```

after:
```
[2023-06-16T06:58:10Z INFO  poseidon_circuit::hash] hash table parallel version took 445.565348ms
[2023-06-16T07:03:29Z INFO  poseidon_circuit::hash] final state took 110.563839ms
```
around 556ms