@noel2004 2023-06-12
> @noel2004 Is it true that each `((inp, control), check)` occupies **exactly one** row in the `fill_hash_tbl_body` region?
> 
> https://github.com/scroll-tech/poseidon-circuit/blob/89e9c0456f0fa00d49cd7920c6b437c66b48db17/src/hash.rs#L507

Yes. Each entry occupies one row.