@alannotnerd 2023-06-13
```
[2023-06-13T11:42:40Z INFO  halo2_proofs::circuit::floor_planner::single_pass] 25 sub_regions of hash table 2nd pass synthesis took 966.003307ms
[2023-06-13T11:42:40Z INFO  poseidon_circuit::hash] hash table parallel version took 972.602083ms
[2023-06-13T11:42:40Z INFO  halo2_proofs::circuit::floor_planner::single_pass] 24 sub_regions of permute state 2nd pass synthesis took 41.37005ms
[2023-06-13T11:42:40Z INFO  poseidon_circuit::hash] final state took 834.965283ms
```