@kunxian-xia 2023-06-18
We got the following stats when testing in scroll-zkevm (tested on a machine with 128 threads):
```
[2023-06-16T06:58:10Z INFO  poseidon_circuit::hash] hash table parallel version took 445.565348ms
[2023-06-16T07:00:47Z INFO  poseidon_circuit::hash] hash table parallel version took 452.829501ms
[2023-06-16T07:03:29Z INFO  poseidon_circuit::hash] hash table parallel version took 490.576667ms
[2023-06-16T06:58:10Z INFO  poseidon_circuit::hash] final state took 105.097895ms
[2023-06-16T07:00:47Z INFO  poseidon_circuit::hash] final state took 103.493175ms
[2023-06-16T07:03:29Z INFO  poseidon_circuit::hash] final state took 110.563839ms
```